create root component for data adapter
